+++ handlers/product-move-api/cfn.yaml
@@ -64,6 +64,8 @@ Resources:
           Resource:
             - arn:aws:sqs:eu-west-1:865473395570:contributions-thanks
             - arn:aws:sqs:eu-west-1:865473395570:contributions-thanks-dev
+            - !GetAtt RefundQueue.Arn
+            - !GetAtt SalesforceTrackingQueue.Arn
       - AWSLambdaBasicExecutionRole
       - Statement:
           Effect: Allow
