+++ handlers/product-move-api/src/test/scala/com/gu/productmove/zuora/MockSQS.scala
@@ -1,6 +1,7 @@
 package com.gu.productmove.zuora
 
 import com.gu.productmove.refund.RefundInput
+import com.gu.productmove.salesforce.Salesforce.SalesforceRecordInput
 import com.gu.productmove.{EmailMessage, SQS}
 import com.gu.productmove.zuora.GetSubscription
 import com.gu.productmove.zuora.GetSubscription.GetSubscriptionResponse
@@ -8,9 +9,9 @@ import com.gu.productmove.zuora.CreateSubscriptionResponse
 import software.amazon.awssdk.services.sqs.SqsAsyncClient
 import zio.{IO, ZIO}
 
-class MockSQS(responses: Map[EmailMessage | RefundInput, Unit]) extends SQS {
+class MockSQS(responses: Map[EmailMessage | RefundInput | SalesforceRecordInput, Unit]) extends SQS {
 
-  private var mutableStore: List[EmailMessage | RefundInput] = Nil // we need to remember the side effects
+  private var mutableStore: List[EmailMessage | RefundInput | SalesforceRecordInput] = Nil // we need to remember the side effects
 
   def requests = mutableStore.reverse
 
@@ -29,8 +30,16 @@ class MockSQS(responses: Map[EmailMessage | RefundInput, Unit]) extends SQS {
       case Some(stubbedResponse) => ZIO.succeed(stubbedResponse)
       case None => ZIO.fail(s"wrong input, refund input was $refundInput")
   }
+
+  override def queueSalesforceTracking(salesforceRecordInput: SalesforceRecordInput): ZIO[Any, String, Unit] = {
+    mutableStore = salesforceRecordInput :: mutableStore
+
+    responses.get(salesforceRecordInput) match
+      case Some(stubbedResponse) => ZIO.succeed(stubbedResponse)
+      case None => ZIO.fail(s"wrong input, salesforce record input was $salesforceRecordInput")
+  }
 }
 
 object MockSQS {
-  def requests: ZIO[MockSQS, Nothing, List[EmailMessage | RefundInput]] = ZIO.serviceWith[MockSQS](_.requests)
+  def requests: ZIO[MockSQS, Nothing, List[EmailMessage | RefundInput | SalesforceRecordInput]] = ZIO.serviceWith[MockSQS](_.requests)
 }
